Update fpu.yar
Updated this rule to not use the deprecated version of entrypoint, but to use the recommended approach that yara gives you when you try to run this rule as-is. I do not have a valid sample to test this on, so I cannot confirm whether this breaks detection, but yara no longer gives the warning with these changes in place.